refactor: use Output enum instead of trait generics
