+++ src/formats.rs
@@ -1,43 +1,42 @@
-use crate::io::RdfIO;
+use crate::converter::RdfIO;
+use crate::io::Output;
 use sophia::turtle::parser::nt::NTriplesParser;
 use sophia::turtle::parser::turtle::TurtleParser;
 use sophia::turtle::serializer::nt::NtSerializer;
 use sophia::turtle::serializer::turtle::TurtleSerializer;
 use sophia::xml::parser::RdfXmlParser;
 use sophia::xml::serializer::RdfXmlSerializer;
-use std::fs::File;
-use std::io::{BufWriter, Write};
 
 pub(crate) struct NTriples;
 pub(crate) struct Turtle;
 pub(crate) struct RdfXml;
 
-impl<'a, W: Write> RdfIO<'a, W, NTriplesParser, NtSerializer<W>> for NTriples {
+impl<'a> RdfIO<'a, NTriplesParser, NtSerializer<Output>> for NTriples {
     fn parser(&self) -> NTriplesParser {
         NTriplesParser {}
     }
 
-    fn serializer(&self, writer: W) -> NtSerializer<W> {
+    fn serializer(&self, writer: Output) -> NtSerializer<Output> {
         NtSerializer::new(writer)
     }
 }
 
-impl<'a, W: Write> RdfIO<'a, W, TurtleParser, TurtleSerializer<W>> for Turtle {
+impl<'a> RdfIO<'a, TurtleParser, TurtleSerializer<Output>> for Turtle {
     fn parser(&self) -> TurtleParser {
         TurtleParser { base: None }
     }
 
-    fn serializer(&self, writer: W) -> TurtleSerializer<W> {
+    fn serializer(&self, writer: Output) -> TurtleSerializer<Output> {
         TurtleSerializer::new(writer)
     }
 }
 
-impl<'a, W: Write> RdfIO<'a, W, RdfXmlParser, RdfXmlSerializer<W>> for RdfXml {
+impl<'a> RdfIO<'a, RdfXmlParser, RdfXmlSerializer<Output>> for RdfXml {
     fn parser(&self) -> RdfXmlParser {
         RdfXmlParser { base: None }
     }
 
-    fn serializer(&self, writer: W) -> RdfXmlSerializer<W> {
+    fn serializer(&self, writer: Output) -> RdfXmlSerializer<Output> {
         RdfXmlSerializer::new(writer)
     }
 }
